Fixed or removed broken invariants
NoDirtyReads was faulty as it was based on the assumption that once a ledger was closed, no further writes to bookies was possible. However, when multiple recovery clients are active, it is possible for one to close the ledger while another is writing back additional entries, advancing the LAC on one or more bookies. This is ok as the ledger metadata defines the entry range of a ledger, not the bookies. Therefore I removed this invariant.

A similar problem afflicted NoDivergenceBetweenClientAndMetaData but the solution was to ensure that the LAC rule applied only to non-recovery clients.